Finished GUI for report time. Also, timeregistration uses float now

+++ test/dtu/planning/acceptance_tests/CorrectReportedTimeSteps.java
@@ -35,7 +35,7 @@ public class CorrectReportedTimeSteps {
 
 	private TimeRegistration time;
 	//private TimeRegistration timereg;
-	private int timeregOld;
+	private float timeregOld;
 
 	private String activityName;
 	private GregorianCalendar date;
